+++ FrontEnd/src/components/HeaderFooter/header/navbar/Profile.module.css
@@ -10,81 +10,6 @@
   align-items: flex-end;
 
   border-radius: 100px;
-  border: 1px solid #e2e5eb;
-  background: #fff;
-}
-
-.header-profile-tab {
-  display: flex;
-  padding: 5px 0px;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-}
-
-.header-profile-tab__text {
-  color: #3c4044;
-  text-align: center;
-  font-feature-settings: "calt" off;
-  text-decoration: none;
-
-  font-family: Inter, sans-serif;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-  letter-spacing: -0.14px;
-}
-
-.header-profile-dropdown-menu {
-  display: flex;
-  width: 160px;
-  padding: 4px 0px;
-  flex-direction: column;
-  align-items: flex-start;
-
-  position: absolute;
-  z-index: 1;
-  top: 44px;
-  left: 1321px;
-  border-radius: 2px;
-  background: var(--conditional-pop-over, #fff);
-  /* drop-shadow/0.12+0.8+0.5 */
-  box-shadow: 0px 9px 28px 8px rgba(0, 0, 0, 0.05),
-    0px 6px 16px 0px rgba(0, 0, 0, 0.08), 0px 3px 6px -4px rgba(0, 0, 0, 0.12);
-}
-
-.header-profile-dropdown-menu button,
-.header-profile-dropdown-menu button:focus,
-.header-profile-dropdown-menu button:active {
-  outline: none;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-}
-
-.header-profile-dropdown-menu__text,
-.header-profile-dropdown-menu button {
-  display: flex;
-  padding: 5px 12px;
-  align-items: center;
-  gap: 8px;
-  align-self: stretch;
-
-  color: #3c4044;
-  text-align: center;
-  font-feature-settings: "calt" off;
-  text-decoration: none;
-
-  font-family: Inter, sans-serif;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-  letter-spacing: -0.14px;
-}
-
-.header-profile-dropdown-menu__text:focus {
-  background: var(--main-grey-10, #eeeff1);
-  cursor: pointer;
+  border: 1px solid var(--main-black-20, #E2E5EB);
+  background: var(--conditional-pop-over, #FFF)
 }
